+++ package.json
@@ -112,6 +112,7 @@
     "minimatch@3.1.2": "3.1.5",
     "serialize-javascript": "^7.0.5",
     "socks": ">=2.8.3",
+    "vite-plugin-node-polyfills": "patch:vite-plugin-node-polyfills@npm%3A0.28.0#~/.yarn/patches/vite-plugin-node-polyfills-npm-0.28.0-bac39616ee.patch",
     "webpackbar": "^7.0.0"
   },
   "browserslist": [
