+++ packages/api/index.browser.ts
@@ -1,7 +1,13 @@
 import { startBackendWorker } from '@actual-app/core/platform/client/backend-worker';
 import { send } from '@actual-app/core/platform/client/connection';
 import type { InitConfig } from '@actual-app/core/server/main';
-import workerCode from 'virtual:actual-worker-code';
+
+// Vite inlines the worker (its wasm + filesystem data are embedded too) as a
+// Blob URL. This is the point of the browser build: consumer bundlers never see
+// a worker entry to re-bundle, and no asset files are fetched, so
+// `import '@actual-app/api'` + `init()` works in any bundler with no config
+// (only COOP/COEP headers are required, for SharedArrayBuffer).
+import InlineWorker from './browser-worker?worker&inline';
 
 export * from './methods';
 export * as utils from './utils';
@@ -11,19 +17,7 @@ let worker: Worker | null = null;
 export async function init(
   config: InitConfig = {},
 ): Promise<{ send: typeof send }> {
-  // The worker is inlined (its wasm + filesystem data are embedded too) and
-  // spawned from a Blob URL. This is the point of the browser build: consumer
-  // bundlers never see a worker entry to re-bundle, and no asset files are
-  // fetched, so `import '@actual-app/api'` + `init()` works in any bundler with
-  // no configuration (only COOP/COEP headers are required, for SharedArrayBuffer).
-  const blobUrl = URL.createObjectURL(
-    new Blob([workerCode], { type: 'text/javascript' }),
-  );
-  try {
-    worker = new Worker(blobUrl);
-  } finally {
-    URL.revokeObjectURL(blobUrl);
-  }
+  worker = new InlineWorker();
 
   try {
     await startBackendWorker(worker, config);
