+++ packages/api/package.json
@@ -49,7 +49,7 @@
     }
   },
   "scripts": {
-    "build": "vite build && tsgo --emitDeclarationOnly && vite build --config vite.browser-worker.config.mts && vite build --config vite.browser.config.mts",
+    "build": "vite build && tsgo --emitDeclarationOnly && vite build --config vite.browser.config.mts",
     "test": "vitest --run",
     "e2e": "playwright test",
     "typecheck": "tsgo -b && tsc-strict"
@@ -68,7 +68,7 @@
     "rollup-plugin-visualizer": "^7.0.1",
     "typescript-strict-plugin": "^2.4.4",
     "vite": "^8.0.16",
-    "vite-plugin-node-polyfills": "^0.28.0",
+    "vite-plugin-node-polyfills": "patch:vite-plugin-node-polyfills@npm%3A0.28.0#~/.yarn/patches/vite-plugin-node-polyfills-npm-0.28.0-bac39616ee.patch",
     "vitest": "^4.1.8"
   },
   "engines": {
